requirements.txt
alabaster==1.0.0 # to modify in case we want a more general code
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asciitree==0.3.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
bokeh==3.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
codeium-jupyter==1.1.21
comm==0.2.2
contourpy==1.3.3
crc32c==2.4
cycler==0.12.1
debugpy @ file:///home/task_176242086745011/conda-bld/debugpy_1762421636577/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
defusedxml==0.7.1
dicomweb-client==0.59.3
docutils==0.21.2
et_xmlfile==2.0.0
eval_type_backport==0.3.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fasteners==0.20
fastjsonschema==2.21.2
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.61.0
fqdn==1.5.1
fsspec==2024.12.0
Glymur==0.13.6
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.2
idna==3.11
imagecodecs==2024.9.22
ImageIO==2.37.2
imagesize==1.4.1
imbalanced-learn==0.14.0
imblearn==0.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
iniconfig==2.3.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipykernel_1761567932/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1764766793/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
llvmlite==0.43.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.9.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mistune==3.1.4
-e git+https://github.com/aymenSadraoui/ML-Model-Integrating-Computational-Pathology-to-Predict-Early-Recurrence-of-HCC.git@d6248b68a89344ada38c45be2e6c429c7fdf3bd3#egg=ML_HCC_Recurrence_Prediction
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.6
notebook==7.1.3
notebook_shim==0.2.4
numba==0.60.0
numcodecs==0.16.5
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
openslide-python==1.4.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.2.3
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pillow==11.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
psutil @ file:///home/task_176189642211680/conda-bld/psutil_1761896463664/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydicom==3.0.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pynndescent==0.5.13
pyparsing==3.2.5
pytest==9.0.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.2
pyzmq @ file:///home/task_176237559504765/conda-bld/pyzmq_1762375608075/work
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.37.0
requests==2.32.3
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.0.6
shellingham==1.5.4
SimpleITK==2.4.0
simsimd==6.5.3
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
snowballstemmer==3.0.1
soupsieve==2.8
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
stringzilla==4.4.0
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tiatoolbox==1.6.0
tifffile==2024.9.20
timm==1.0.12
tinycss2==1.4.0
torch==2.5.1
torchvision==0.20.1
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
triton==3.1.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
umap-learn==0.5.7
universal_pathlib==0.2.6
uri-template==1.3.0
urllib3==2.6.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
wheel==0.45.1
widgetsnbextension==4.0.10
wsidicom==0.22.0
xyzservices==2025.11.0
zarr==2.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_zipp_1764460141/work
